user-service/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
motor
pymongo
python-dotenv
email-validator
bcrypt
